Implemeting main shoppingcard methods, and checkout without paiment

+++ database/migrations/2020_11_16_103316_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->decimal('price');
+            $table->bigInteger('price');
             $table->string('image');
             $table->string('description');
             $table->timestamps();
